script/requirements.txt
azure-identity
azure-mgmt-web
python-dotenv